+++ docs/src/content/blog/the-fresh-keystatic-cli.mdoc
@@ -13,12 +13,12 @@ The freshly baked Keystatic CLI is now published on NPM 🔥
 
 Simply run `npm create @keystatic@latest` in your terminal to try it out!
 
-The CLI creates a basic Next.js app (using the new App Router), and comes with a simple Keystatic integration showing how to use the [Collections API](/docs/collections-and-singletons#collections).
-
-We're also planning to add more frameworks in the near future, so stay tuned!
-
 Check out the video below, where Simon demoes the CLI, spins up the project and adds new a "Draft" checkbox field to the Keystatic config file:
 
 {% embed
    mediaType="video"
    embedCode="<iframe width=\"560\" height=\"315\" src=\"https://www.youtube.com/embed/N3pywDv-0Tg\" title=\"YouTube video player\" frameborder=\"0\" allow=\"accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture; web-share\" allowfullscreen></iframe>" /%}
+
+The CLI creates a basic Next.js app (using the new App Router), and comes with a simple Keystatic integration showing how to use the [Collections API](/docs/collections-and-singletons#collections).
+
+We're also planning to add more frameworks in the near future, so stay tuned!
