+++ docs/src/utils/index.ts
@@ -1,3 +1,4 @@
+import { format, parse } from 'date-fns';
 import { DEFAULT_IMG_WIDTHS } from '../constants';
 
 export function cx(...classes: string[]) {
@@ -14,4 +15,12 @@ export async function copyTextToClipboard(text: string) {
   if ('clipboard' in navigator) {
     return await navigator.clipboard.writeText(text);
   }
+}
+
+export function parseAndFormatPublishedDate(publishedDate: string) {
+  const today = new Date();
+  const parsedDate = parse(publishedDate, 'yyyy-M-d', today);
+  const formattedDate = format(parsedDate, 'MMMM do, yyyy');
+
+  return { parsedDate, formattedDate };
 }
